requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.4
numpy==1.26.4
pyarrow==17.0.0
PyJWT==2.9.0
pylance==0.18.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
uritemplate==4.1.1
